Refactor string cleaning and substitute state rename.
